pyproject.toml

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.2"
tag_format = "v$version"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version",
    "dramatiq_azure/__version__.py",
]

[tool.poetry]
name = "dramatiq_azure"
version = "0.3.2"
description = "Azure Queue Brokers for Dramatiq"
authors = ["Stanislas H.B. Sodonon <stanislas.sodonon@gmail.com>"]
repository = "https://github.com/bidossessi/dramatiq-azure"
readme = 'README.md'
keywords = ['dramatiq', 'azure', 'queue', 'broker']
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
dramatiq = "^2.0.0"
azure-core = "^1.36.0"
azure-storage-queue = "^12.14.1"
azure-identity = {version = "^1.25.1", optional = true}

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^25.1.0"
pytest = "^8.2.1"
pytest-cov = "^6.1.1"
mypy = "^1.11"
pytest-env = "^1.1.3"
commitizen = "^4.7.0"
flake8-bugbear = ">=23.7.10,<25.0.0"
mccabe = "^0.7.0"
isort = ">=5.12,<7.0"
pre-commit = ">=3.8,<5.0"
pytest-mock = "^3.14.0"

[tool.poetry.extras]
identity = ["azure-identity"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.pytest_cache
  | dist
  | build
  | docs
)/
'''
include = '\.pyi?$'
line-length = 80

[tool.isort]
atomic = true
combine_as_imports = true
group_by_package = true
force_alphabetical_sort_within_sections = true
force_grid_wrap = 2
line_length = 80
multi_line_output = 3
profile = "black"
